# naberal_base/api/openapi.yaml
openapi: 3.1.0
info:
  title: KIS Estimator API
  version: 1.0.0
  description: |
    전기 패널 견적 자동화 시스템 - Contract-First + Evidence-Gated + Zero-Mock

    **Phase VII - API 프로브 안정화 완료**

    ## 핵심 원칙
    - Contract-First: 스키마가 코드보다 우선
    - 5xx Fail Fast: 재시도 없이 즉시 반환
    - Evidence-Gated: 모든 응답에 traceId 포함

    ## API 엔드포인트 (20개)

    ### 헬스체크/프로브 (4개)
    - GET /health - 전체 헬스체크 (DB + Cache + SSOT)
    - GET /health/live - Liveness probe (< 50ms)
    - GET /health/db - Database health probe
    - GET /readyz - Readiness check (DB + Redis + K-PEW)

    ### 카탈로그 (7개)
    - GET /v1/catalog/items - 카탈로그 목록 조회 (페이지네이션)
    - GET /v1/catalog/items/{sku} - 단일 항목 조회
    - GET /v1/catalog/stats - 통계
    - **GET /v1/catalog/metrics - 성능 메트릭 (Phase VII)**
    - **POST /v1/catalog/cache/invalidate - 캐시 무효화 (Phase VII)**
    - **POST /v1/catalog/cache/warm - 캐시 워밍 (Phase VII)**

    ### 견적 (4개)
    - POST /v1/estimate [DEPRECATED] - 견적 생성 (backward compatibility)
    - POST /v1/estimate/plan - EPDL 계획 생성
    - POST /v1/estimate/execute - EPDL 실행
    - GET /v1/estimate/{estimate_id} - 견적 조회

    ### 견적서 (5개 - Phase X/XI/XII)
    - **POST /v1/quotes - 견적서 생성 (SSOT 검증)**
    - **GET /v1/quotes/{quote_id} - 견적서 조회**
    - **POST /v1/quotes/{quote_id}/approve - 견적서 승인 (RBAC)**
    - **POST /v1/quotes/{quote_id}/pdf - PDF 생성 + S3 아카이빙**
    - **GET /v1/quotes/{quote_id}/pdf/url - Pre-signed URL 조회**
  contact:
    name: NABERAL KIS Estimator Team
    email: support@kis-estimator.com
...
servers:
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: string
          description: "에러 코드 (예: CAT-VAL, DB_INSERT_ERROR)"
          example: "CAT-VAL"
        message:
          type: string
          description: 에러 메시지
          example: "Invalid kind: invalid_value"
        hint:
          type: string
          description: 해결 힌트 (선택사항)
          example: "Valid kinds: ['accessory', 'breaker', 'busbar', 'cable', 'enclosure', 'misc']"
        traceId:
          type: string
          format: uuid
          description: 추적 ID (Evidence 연동)
          example: "550e8400-e29b-41d4-a716-446655440000"
        meta:
          type: object
          description: 메타데이터
          properties:
            dedupKey:
              type: string
              description: 중복 방지 키
              example: "dedup-1234567890"
            timestamp:
              type: string
              format: date-time
              description: 에러 발생 시간
...
    BreakerSpec:
      type: object
      required:
        - poles
        - current
        - frame
      properties:
        type:
          type: string
          enum: [MCCB, ELB]
          description: 차단기 종류 (MCCB=배선용, ELB=누전)
          example: "MCCB"
        series:
          type: string
          enum: [경제형, 표준형, 소형]
          description: 시리즈 (경제형=37kA, 표준형=50kA, 소형=2.5kA)
          example: "경제형"
        poles:
          type: integer
          minimum: 2
          maximum: 4
          description: 극수 (2/3/4)
          example: 4
        current:
          type: integer
          minimum: 1
          maximum: 1000
          description: 전류 (A)
          example: 75
        frame:
          type: integer
          minimum: 1
          maximum: 1000
          description: 프레임 (AF)
          example: 100
      description: |
        차단기 스펙

        **비즈니스 규칙**: frame >= current (프레임은 전류 이상이어야 함)
        **기본 원칙**: 특별 요청 없으면 경제형 사용 (경제형 37kA, 표준형 50kA)
...
    EnclosureSpec:
      type: object
      required:
        - type
        - material
      properties:
        type:
          type: string
          enum: [옥내노출, 옥외노출, 옥내자립, 옥외자립, 전주부착형, FRP함, 하이박스, 매입함]
          description: 외함 종류
          example: "옥내노출"
        material:
          type: string
          enum: [STEEL_1.6T, STEEL_1.0T, SUS201_1.2T, SUS304_1.2T]
          description: 재질/두께
          example: "STEEL_1.6T"
      description: |
        외함 스펙

        **베이스 규칙** (옥내자립, 옥외자립, FRP함):
        - 크기: W=외함W, H=200mm, D=외함D
        - 가격: ((W*H)/90000)*32000원

        **전주부착형 특수 규칙**:
        - 반드시 옥외함
        - 브라켓 용접비: +20,000원
    PlanRequest:
      type: object
      required:
        - customer_name
        - project_name
        - main_breaker
        - branch_breakers
      properties:
        customer_name:
          type: string
          minLength: 1
          description: 고객사명
          example: "한국전기주식회사"
        project_name:
          type: string
          minLength: 1
          description: 프로젝트명
          example: "서울 본사 전기설비 공사"
        panel_name:
          type: string
          description: 분전반명 (없으면 "분전반")
          default: "분전반"
          example: "MCC-1"
        enclosure:
          $ref: '#/components/schemas/EnclosureSpec'
        breaker_brand:
          type: string
          enum: [상도차단기, LS차단기, 한국산업]
          description: 차단기 브랜드
          default: "상도차단기"
          example: "상도차단기"
        main_breaker:
          $ref: '#/components/schemas/BreakerSpec'
        branch_breakers:
          type: array
          items:
            $ref: '#/components/schemas/BreakerSpec'
          description: 분기 차단기 목록
        accessories:
          type: array
          items:
            $ref: '#/components/schemas/AccessorySpec'
          description: 부속자재 목록
          default: []
    PlanResponse:
      type: object
      required:
        - plan_id
        - estimate_id
        - verb_specs
        - specs_count
        - is_valid
        - created_at
      properties:
        plan_id:
          type: string
          pattern: '^EST-\\d{20}$'
          description: 견적 ID (EST-YYYYMMDDHHMMSSffffff)
          example: "EST-20251025123456789012"
        estimate_id:
          type: string
          description: 견적 ID (plan_id와 동일)
          example: "EST-20251025123456789012"
        verb_specs:
          type: array
          items:
            type: object
          description: 생성된 VerbSpec 리스트
        specs_count:
          type: integer
          description: VerbSpec 개수
          example: 8
        is_valid:
          type: boolean
          description: 스키마 검증 통과 여부
          example: true
        created_at:
          type: string
          format: date-time
          description: 생성 시간 (ISO 8601)
...
    EstimateDetailResponse:
      type: object
      required:
        - estimate_id
        - plan
        - execution_history
      properties:
        estimate_id:
          type: string
          description: 견적 ID
          example: "EST-20251025123456789012"
        plan:
          type: object
          description: 계획 상세
        created_at:
          type: string
          format: date-time
          description: 생성 시간
        execution_history:
          type: array
          items:
            type: object
          description: 실행 이력

    # ========== 성능 메트릭 스키마 (Phase VII) ==========
...
    CreateQuoteRequest:
      type: object
      required:
        - items
        - client
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuoteItemRequest'
          description: 견적 항목 (최소 1개 필수)
        client:
          type: string
          minLength: 1
          description: 고객사명
          example: "ABC건설"
        terms_ref:
          type: string
          description: "결제 조건 (예: NET30, NET60)"
          example: "NET30"
...
    CreateQuoteResponse:
      type: object
      required:
        - quote_id
        - totals
        - approval_required
        - evidence_hash
        - created_at
      properties:
        quote_id:
          type: string
          description: 견적 ID
          example: "QT-20251118-0001"
        totals:
          $ref: '#/components/schemas/TotalsResponse'
        approval_required:
          type: boolean
          description: 승인 필요 여부 (≥50,000,000 KRW)
          example: false
        evidence_hash:
          type: string
          description: Evidence 해시
          example: "a3b2c1d4e5f6..."
        created_at:
          type: string
          format: date-time
          description: 생성 시간
    QuoteResponse:
      type: object
      required:
        - quote_id
        - items
        - client
        - totals
        - status
        - evidence_hash
        - created_at
        - updated_at
      properties:
        quote_id:
          type: string
          description: 견적 ID
          example: "QT-20251118-0001"
        items:
          type: array
          items:
            type: object
          description: 견적 항목 목록
        client:
          type: string
          description: 고객사명
          example: "ABC건설"
        terms_ref:
          type: string
          description: 결제 조건
          example: "NET30"
        totals:
          $ref: '#/components/schemas/TotalsResponse'
        status:
          type: string
          enum: [DRAFT, APPROVED]
          description: 상태
          example: "DRAFT"
        evidence_hash:
          type: string
          description: Evidence 해시
        created_at:
          type: string
          format: date-time
          description: 생성 시간
        updated_at:
          type: string
          format: date-time
          description: 수정 시간
        approved_at:
          type: string
          format: date-time
          description: 승인 시간 (승인된 경우)
        approved_by:
          type: string
          description: 승인자 (승인된 경우)
    ApprovalRequest:
      type: object
      required:
        - actor
      properties:
        actor:
          type: string
          minLength: 1
          description: 승인자 식별자 (이메일 또는 사용자명)
          example: "admin@company.com"
        comment:
          type: string
          description: 승인 코멘트 (선택)
          example: "승인합니다"
    ApprovalResponse:
      type: object
      required:
        - quote_id
        - status
        - approved_at
        - approved_by
        - evidence_entry
      properties:
        quote_id:
          type: string
          description: 견적 ID
          example: "QT-20251118-0001"
        status:
          type: string
          enum: [APPROVED]
          description: 상태
          example: "APPROVED"
        approved_at:
          type: string
          format: date-time
          description: 승인 시간
        approved_by:
          type: string
          description: 승인자
          example: "admin@company.com"
        evidence_entry:
          type: string
          description: Evidence 로그 엔트리
          example: "audit-log-12345"
    PDFGenerateResponse:
      type: object
      required:
        - id
        - pdf_path
        - approved
        - audit_passed
      properties:
        id:
          type: string
          description: 견적 ID
          example: "QT-20251118-0001"
        pdf_path:
          type: string
          description: PDF 파일 경로
          example: "/out/quotes/QT-20251118-0001.pdf"
        evidence_hash:
          type: string
          description: Evidence 해시
          example: "a3b2c1d4e5f6..."
        approved:
          type: boolean
          description: 승인 여부
          example: true
        audit_passed:
          type: boolean
          description: PDF 감사 통과 여부
          example: true
        s3_url:
          type: string
          format: uri
          description: S3 업로드 URL (업로드된 경우)
          example: "https://s3.amazonaws.com/bucket/quotes/QT-20251118-0001.pdf"
    PresignedURLResponse:
      type: object
      required:
        - url
        - expires_at
        - approved
        - evidence_hash
        - storage_mode
      properties:
        url:
          type: string
          format: uri
          description: "Pre-signed URL 또는 local:// URL"
          example: "https://s3.amazonaws.com/bucket/quotes/QT-20251118-0001.pdf?signature=..."
        expires_at:
          type: string
          format: date-time
          description: URL 만료 시간 (15분)
        approved:
          type: boolean
          description: 승인 여부
          example: true
        evidence_hash:
          type: string
          description: Evidence 해시
        storage_mode:
          type: string
          enum: [s3, local]
          description: 저장소 모드
          example: "s3"
paths:
...
  /v1/estimate:
    post:
      tags:
        - estimate
      summary: '[DEPRECATED] 견적 생성'
      description: |
        ⚠️ **DEPRECATED**: 이 엔드포인트는 deprecated되었습니다.

        **대신 사용**: POST /v1/estimate/plan

        **Sunset Date**: 90일 후

        **Deprecation Headers**:
        - Deprecation: true
        - Sunset: {sunset_date}
        - Link: </v1/estimate/plan>; rel="alternate"

        이 엔드포인트는 내부적으로 /v1/estimate/plan에 위임합니다.
      operationId: createEstimateDeprecated
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
      responses:
        '201':
          description: 계획 생성 성공
          headers:
            Deprecation:
              schema:
                type: string
                enum: [true]
              description: Deprecated 플래그
            Sunset:
              schema:
                type: string
                format: date-time
              description: Sunset 날짜
            Link:
              schema:
                type: string
              description: 대체 엔드포인트 링크
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: 유효하지 않은 요청 (VerbSpec 생성/검증 실패)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 내부 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/estimate/{estimate_id}:
    get:
      tags:
        - estimate
      summary: 견적 조회
      description: |
        estimate_id로 견적 상세 조회

        **응답 포함**:
        - plan: VerbSpec 계획
        - execution_history: 실행 이력 (있는 경우)
      operationId: getEstimate
      parameters:
        - name: estimate_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^EST-\\d{20}$'
          description: 견적 ID
          example: "EST-20251025123456789012"
      responses:
        '200':
          description: 견적 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateDetailResponse'
        '404':
          description: 견적을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 데이터베이스 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== Phase VII Catalog Endpoints ==========
...
  /v1/quotes:
    post:
      tags:
        - quotes
      summary: 견적서 생성
      description: |
        견적서 생성 - Phase X (SSOT 검증 포함)

        **SSOT 검증**:
        - UOM: uom.json 기준
        - Discount Rules: discount_rules.json 기준
        - Rounding: HALF_EVEN 방식

        **초기 상태**: DRAFT
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: 견적서 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateQuoteResponse'
        '400':
          description: 잘못된 요청 (SSOT 검증 실패)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "INVALID_UOM"
                message: "Invalid unit of measure: INVALID"
                hint: "Valid UOMs: EA, SET, M, KG, PCS, BOX, ROLL"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: 데이터베이스 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/quotes/{quote_id}:
    get:
      tags:
        - quotes
      summary: 견적서 조회
      description: |
        quote_id로 견적서 상세 조회 - Phase X

        **응답 포함**:
        - items: 견적 항목 리스트
        - totals: 합계 (subtotal, tax, total)
        - status: DRAFT | APPROVED
        - created_at, updated_at
      operationId: getQuote
      parameters:
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 견적서 ID (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: 견적서 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '404':
          description: 견적서를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 데이터베이스 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/quotes/{quote_id}/approve:
    post:
      tags:
        - quotes
      summary: 견적서 승인
      description: |
        견적서 승인 - Phase XI (RBAC: APPROVER 권한 필요)

        **상태 전환**: DRAFT → APPROVED
        **권한**: APPROVER 역할 필수
        **부작용**: 승인 후 수정 불가
      operationId: approveQuote
      parameters:
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 견적서 ID (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: 승인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
        '403':
          description: 권한 없음 (APPROVER 역할 필요)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "FORBIDDEN"
                message: "User lacks APPROVER role"
                hint: "Contact administrator for role assignment"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '404':
          description: 견적서를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 이미 승인된 견적서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 데이터베이스 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/quotes/{quote_id}/pdf:
    post:
      tags:
        - quotes
      summary: PDF 생성
      description: |
        견적서 PDF 생성 - Phase XII

        **PDF 표준**:
        - A4 사이즈 (210×297mm)
        - 여백: 20mm (상하좌우)
        - Evidence Footer: SHA256 해시, 생성 시각 포함
        - 한글 폰트: NanumGothic

        **S3 아카이빙**:
        - 성공 시: S3 URL 반환
        - 실패 시: Graceful degradation (X-Archive-Warn 헤더)
      operationId: generateQuotePDF
      parameters:
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 견적서 ID (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '201':
          description: PDF 생성 성공
          headers:
            X-Archive-Warn:
              schema:
                type: string
              description: S3 아카이빙 실패 경고 (선택적)
              example: "S3 archiving failed - PDF available temporarily"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PDFGenerateResponse'
        '404':
          description: 견적서를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: PDF 생성 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/quotes/{quote_id}/pdf/url:
    get:
      tags:
        - quotes
      summary: PDF Pre-signed URL 조회
      description: |
        S3 Pre-signed URL 조회 - Phase XII

        **만료 시간**: 15분
        **용도**: 클라이언트 PDF 다운로드
        **Graceful Degradation**: S3 미사용 시 404 반환
      operationId: getQuotePDFUrl
      parameters:
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 견적서 ID (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Pre-signed URL 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedURLResponse'
        '404':
          description: PDF를 찾을 수 없음 또는 S3 미사용
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: S3 URL 생성 에러 (5xx Fail Fast)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'